+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for AppRole auth method
 - Support for Userpass auth method
 - Generic login method for clients to easily obtain tokens
+- List method for listing available auth methods on a server
 
 ## [0.3.0] - 2021-08-30